+++ glfw/src/win32_platform.h
@@ -500,8 +500,8 @@ void _glfwWaitEventsWin32(void);
 void _glfwWaitEventsTimeoutWin32(double timeout);
 void _glfwPostEmptyEventWin32(void);
 
-void _glfwGetCursorPosWin32(_GLFWwindow* window, double* xpos, double* ypos);
-void _glfwSetCursorPosWin32(_GLFWwindow* window, double xpos, double ypos);
+void _glfwGetCursorPosWin32(_GLFWwindow* window, float* xpos, float* ypos);
+void _glfwSetCursorPosWin32(_GLFWwindow* window, float xpos, float ypos);
 void _glfwSetCursorModeWin32(_GLFWwindow* window, int mode);
 int _glfwGetKeyScancodeWin32(int key);
 GLFWbool _glfwCreateCursorWin32(_GLFWcursor* cursor, const GLFWimage* image, int xhot, int yhot);
